Initial registrar implementation
I'm willing to bet that I've broken the unit tests again. And there's
new functionality that probably doesn't work and needs to be tested.

But I think this is a decent first idea of what I really have to
have. And, if nothing else, it compiles. (Yes, it's pretty much totally
untested at the REPL, also).