Broken: start switching to single server socket
At this rate I'll wind up turning full circle and end up with my 3
original sockets.

This version doesn't compile, and it's really just the first step in
this direction anyway. Need to make the EventPairInterface its own
top-level Component and invert the current dependency that Plugin
Manager has on it.

But I'm out of time for the moment