Debugging startup/reset problems
This version started working better than the previous, after I commented
out some of the "less-important" Components from the System. No idea
whether that fixed the problem, it was an unintended consequence of some
of the refactoring I've been doing on component-dsl to try to help
locate the problem, or if an issue with cljeromq and spec generation has
decided to start causing problems before the problem I've been having
about needing to recurse into nested constructors again.

But I'm out of time for poking at it for today.